dynamic values wip: refactor to new decorator
